SUBMARINE-195. Fix the incorrect project name in merge_submarine_pr.py
### What is this PR for?
When it merges the PR and the committer chose to close Jira, it fails with something like project "HADOOP-SUBMARINE" not be found. Tested that change it to "SUBMARINE" will work.

### What type of PR is it?
Bug Fix
### What is the Jira issue?
https://issues.apache.org/jira/browse/SUBMARINE-195

### How should this be tested?
use the cicd image to do the PR merge. And also choose yes to update the associated JIRA.
It will work with this fix.

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Zhankun Tang <ztang@apache.org>

Closes #7 from tangzhankun/submarine-195 and squashes the following commits:

ca52b46 [Zhankun Tang] SUBMARINE-195. Fix the incorrect project name in merge_submarine_pr.py