SUBMARINE-256. Interpreter support cluster mode
### What is this PR for?
After the interpreter integrates the cluster module,
When the interpreter is started, the `interpreter ID`, `IP` and `port` of the interpreter are registered in the metadata of the cluster, so that the interpreter can be found in the entire cluster and used by the notebook and workbench server.

### What type of PR is it?
[Feature]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/SUBMARINE-256

### How should this be tested?
* [CI Pass](https://travis-ci.org/liuxunorg/hadoop-submarine/builds/601138608)
* InterpreterClusterTest.java

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Xun Liu <liuxun@apache.org>

Closes #60 from liuxunorg/SUBMARINE-256 and squashes the following commits:

dc1ba6e [Xun Liu] SUBMARINE-256. Interpreter support cluster mode