SUBMARINE-295. Modify mysql database submarineDB to submarine
### What is this PR for?
Because the mysql database name is case sensitive,
Now submarine has 2 mysql databases. They are `submarineDB` and `submarineDB_test`.
It is easy to write `DB` as `db` in program development, Causes a program error and is difficult to troubleshoot.
So modify the `submarineDB` and `submarineDB_test` database to `submarine` and `submarine_test`.

### What type of PR is it?
Improvement

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/SUBMARINE-295

### How should this be tested?
* [CI Pass](https://travis-ci.org/liuxunorg/submarine/builds/614558566)

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Xun Liu <liuxun@apache.org>

Closes #95 from liuxunorg/SUBMARINE-295 and squashes the following commits:

549c12a [Xun Liu] SUBMARINE-295. Modify mysql database submarineDB to submarine