SUBMARINE-373. Upgrade cluster module dependent library
### What is this PR for?
The atomix raft library used by the submarine cluster module is version 3.0,
Need to know the list of all servers ip in the cluster to create a cluster.

In the k8s environment, since the IP of the POD is not fixed, we need to upgrade to the atomix-3.1.0 + version and create a cluster through broadcast discovery.

### What type of PR is it?
[Refactoring]

### Todos
* [ ] - Create a cluster through broadcast discovery.

### What is the Jira issue?
* https://issues.apache.org/jira/browse/SUBMARINE-373

### How should this be tested?
* https://travis-ci.org/liuxunorg/submarine/builds/645828692

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Xun Liu <liuxun@apache.org>

Closes #172 from liuxunorg/SUBMARINE-373 and squashes the following commits:

55addda [Xun Liu] relocation
43c3667 [Xun Liu] fixed exclude netty problem
197c86f [Xun Liu] SUBMARINE-373. Upgrade cluster module dependent library version