SUBMARINE-380. Optimize create kind cluster scripts
### What is this PR for?
In the submarine-cloud script for creating a kind cluster,
If the user does not have kind bin installed, after the automatic download, the user is required to manually set the kind to the PATH environment variable.

We need to provide a kind script in the submarine project. This script proxy all kind bin commands. If the kind bin file is not installed locally, the kind script will automatically download the kind bin file. All kind bin operations are performed by this kind script, so that the user does not need to set the PATH environment variable.

### What type of PR is it?
Improvement

### Todos
* [[k8s-e2e] Create k8s e2e test module](https://issues.apache.org/jira/browse/SUBMARINE-385)

### What is the Jira issue?
* https://issues.apache.org/jira/browse/SUBMARINE-380

### How should this be tested?
1. execute `submarine-cloud/hack/kind-create-cluster.sh`
2. execute `submarine-cloud/hack/deploy-submarine.sh`
3. open `http://127.0.0.1` in chrome, you can see submarine workbench
4. execute `submarine-cloud/hack/deploy-submarine.sh -u`, uninstall submarine in kind
5. execute `kind delete cluster`

* https://travis-ci.org/liuxunorg/submarine/builds/650709676

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Xun Liu <liuxun@apache.org>

Closes #183 from liuxunorg/SUBMARINE-380 and squashes the following commits:

7725226 [Xun Liu] SUBMARINE-380. Optimize submarine-cloud to create kind cluster script