SUBMARINE-422. Fix refreshing page returns 404 error
### What is this PR for?
Because the workbench is developed using angular, the adjustment of angular WEB pages is completely controlled by the front end, so when you manually refresh a specific page in the browser, the browser will send the request for this page to the back-end service, but the back-end service only In response to API requests, it will cause the front end to display 404.

The solution is to find that not all API requests directly return the content of the index page,
so that the front end will automatically perform correct page routing processing.

### What type of PR is it?
[Bug Fix | Improvement]

### Todos
* [ ] - Task

### What is the Jira issue?
* https://issues.apache.org/jira/browse/SUBMARINE-422

### How should this be tested?
* https://travis-ci.org/github/liuxunorg/submarine/builds/663406363

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Xun Liu <liuxun@apache.org>

Closes #232 from liuxunorg/SUBMARINE-422 and squashes the following commits:

f098039 [Xun Liu] Safe close InputStreamReader.
4f44976 [Xun Liu] SUBMARINE-422. Fix refreshing page returns 404 error