SUBMARINE-456. Fixed workbench-web Angular fails in travis
### What is this PR for?
Hi jojochuang liuxunorg

This PR is trying to fix the e2e angular testing job fails intermittently. The problem is that file `gecko-response.json` is downloaded from api.github.com so that if the jobs make too many requests at a given moment it fails.

I suggest that we can merge this PR and observe the CI jobs for the next few days or other commits before closing the Jira issue ticket.

### What type of PR is it?
[Bug Fix]

### What is the Jira issue?

https://issues.apache.org/jira/browse/SUBMARINE-456

### How should this be tested?

The PR is tested many times on my CI jobs and it builds successfully.

https://travis-ci.com/github/cchung100m/submarine/builds/155547853

### Screenshots (if appropriate)

<img width="652" alt="螢幕快照 2020-03-25 下午11 00 28" src="https://user-images.githubusercontent.com/6762509/77659627-47a44b00-6fb3-11ea-8593-988c15953ff1.png">

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: cchung100m <cchung100m@cs.ccu.edu.tw>

Closes #250 from cchung100m/SUBMARINE-456 and squashes the following commits:

2e7df04 [cchung100m] SUBMARINE-456. submarine workbench-web Angular fails intermittently