SUBMARINE-467. Simplify the submitter configuration
### What is this PR for?
At now the submarine server supports config more submitters, but it's not necessary in current stage. At the same time it also cause trouble to users. So we'd better simplify the configuration to keep only one submitter in server. Considering the project vision, we set the K8s submitter as the default.

### What type of PR is it?
[Improvement]

### Todos

### What is the Jira issue?
https://issues.apache.org/jira/browse/SUBMARINE-467

### How should this be tested?
https://travis-ci.com/github/jiwq/submarine/builds/160762675

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Wanqiang Ji <wanqiang.ji@gmail.com>

Closes #266 from jiwq/SUBMARINE-467 and squashes the following commits:

6586d0c [Wanqiang Ji] SUBMARINE-467. Simplify the submitter configuration