SUBMARINE-442. Support get job's log in submarine-server REST API
### What is this PR for?

Now we have the "jobs" resource in REST which can do CRUD. We also need a "logs" API to get the job's log output. The URI could be "api/v1/logs"
It should accept parameters like "jobid". Initially, the logs could be aggregated logs of all containers.
Streaming is preferred so that the python client can enable a fancy way for the end-user to check logs

### What type of PR is it?
Feature

### Todos
* [x] - get logs so far

### What is the Jira issue?
https://issues.apache.org/jira/projects/SUBMARINE/issues/SUBMARINE-442

### How should this be tested?
Create a job
Visit /api/v1/logs or /api/v1/logs/{jobid} with a browser

### Screenshots (if appropriate)
http://127.0.0.1:8080/api/v1/jobs/logs
![image](https://user-images.githubusercontent.com/19265751/79961593-8a294c80-84b9-11ea-85ef-9367e17fecc9.png)

http://127.0.0.1:8080/api/v1/jobs/logs/job_1587481945001_0001
![image](https://user-images.githubusercontent.com/19265751/79961674-a0370d00-84b9-11ea-908f-b6bdeceeb6eb.png)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: JohnTing <jot.johnting@gmail.com>

Closes #263 from JohnTing/SUBMARINE-442 and squashes the following commits:

7195a77 [JohnTing] test12
1bb806f [JohnTing] test12
319222c [JohnTing] SUBMARINE-442