SUBMARINE-662. Clarify duplicate experiment error
### What is this PR for?
When we send an experiment POST request that has the same name as the previous ones, it will raise an error. However, the error message does not explain the error clearly, it only says that there is a Conflict. (as the following)
```
{
  "status": "CONFLICT",
  "code": 409,
  "success": null,
  "message": "Conflict",
  "result": null,
  "attributes": {}
}
```
To let users understand the cause of the error better, I change the message from "Conflict" to "Duplicated experiment".

```
{
  "status": "CONFLICT",
  "code": 409,
  "success": null,
  "message": "Duplicated Experiment",
  "result": null,
  "attributes": {}
}
 
```

### What type of PR is it?
[Improvement]

### Todos
* [ ] - Task

### What is the Jira issue?
https://issues.apache.org/jira/browse/SUBMARINE-662

### How should this be tested?
https://travis-ci.org/github/ByronHsu/submarine/builds/736283354

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? Yes/No
* Is there breaking changes for older versions? Yes/No
* Does this needs documentation? Yes/No

Author: ByronHsu <byronhsu1230@gmail.com>

Closes #436 from ByronHsu/SUBMARINE-662 and squashes the following commits:

d786490 [ByronHsu] Update submarine-server/server-core/src/main/java/org/apache/submarine/server/rest/ExperimentRestApi.java
062f89d [ByronHsu] Clarify duplicate error