SUBMARINE-882. fix datadictIT
### What is this PR for?
In travis.ci, this IT will have undeterministic behavior, it is because of the original way to realize the waiting is fluent wait, but it lacks of expected condition.
The other reason of unstable behavior of this test is because of the wrong interaction with components.

https://user-images.githubusercontent.com/55401762/124241557-6e561780-db4e-11eb-8ab4-53a04ffc583e.mp4

As you can see, this drop-down list need to be hovered. But this IT clicked it. This might be the source of unstable.
So it need to be modify with the functions that use explicit wait in AbstractSubmarineIT.
Also found that the window need to set larger to prevent element from disappearing in viewpoint, which makes headless driver can not interact.
Note: the default of size of headless window maximize is (600, 800), when driver hover the target out of this area, it will cause MoveTargetOutOfBounds exception.

### What type of PR is it?
[Bug Fix | Refactoring]

### Todos

### What is the Jira issue?

https://issues.apache.org/jira/projects/SUBMARINE/issues/SUBMARINE-882

### How should this be tested?
```bash
# Step1: Run Submarine workbench on 127.0.0.1:8080 by operator
# Step2: Run testcase
cd submarine-cloud-v2

# Usage: ./hack/run_frontend_e2e.sh ${testcase}
# Example:
./hack/run_frontend_e2e.sh datadictIT
```
### Screenshots (if appropriate)

https://user-images.githubusercontent.com/55401762/124000807-8c186500-da06-11eb-8289-faf054a85650.mp4

### Questions:
* Do the license files need updating? No
* Are there breaking changes for older versions? No
* Does this need new documentation? No

Author: noidname01 <tim983649@gmail.com>

Signed-off-by: Kai-Hsun Chen <kaihsun@apache.org>

Closes #627 from noidname01/SUBMARINE-882 and squashes the following commits:

e736ecce [noidname01] add Hover, HoverAndClick
277ff426 [noidname01] set screen bigger
b9f90cb4 [noidname01] add scroll
6fa22aa2 [noidname01] try to fix bug
e64d9ab7 [noidname01] found where unstable-moreBtn
bae44f03 [noidname01] merge with 878
b3c72257 [noidname01] add more comment to increase readability
5d22c2b1 [noidname01] datadictIT refactor
c329d788 [noidname01] datadictIT refactor