SUBMARINE-964. Pull notebook image is very time consuming
### What is this PR for?
Pulling the image of notebook is a quite time consuming process, and it will make user confuse that what the development is waiting for, so we add a new status - Pulling to indicate  the pending is caused by pulling operation.

### What type of PR is it?
Improvement.

### Todos
N/A

### What is the Jira issue?
https://issues.apache.org/jira/browse/SUBMARINE-964
### How should this be tested?
Observe the status in notebook page after creating a notebook.
### Screenshots (if appropriate)
<img width="1346" alt="截圖 2021-09-01 下午6 44 52" src="https://user-images.githubusercontent.com/5687317/131663293-2e7a0e60-6569-4379-bea2-f6f585fb9fb7.png">

### Questions:
* Do the license files need updating? No
* Are there breaking changes for older versions? No
* Does this need new documentation? No

Author: Brandon Lin <fatallin@gmail.com>

Signed-off-by: Kevin <pingsutw@apache.org>

Closes #735 from FatalLin/SUBMARINE-964 and squashes the following commits:

7b9eab1b [Brandon Lin] polish and correct typo
f95373bd [Brandon Lin] code polish
d28bedd0 [Brandon Lin] leverge labelselector to find pods specifically
6df144b2 [Brandon Lin] remove unused import
408d3ab4 [Brandon Lin] add new status pulling for notebook