SUBMARINE-1012. Submarine CR status is inconsistent with actual state
### What is this PR for?
When creating the submarine using the operator, the status of the submarine CR shows mlflow failed due to time-out processing. However, the mlflow deployment is running.

#### Cause:
Submarine operator treat progressing error, ex. ProgressDeadlineExceeded, as a fatal error. However, the deployment controller will continue to process failed deployment, so the deployment may be running in several minutes. This causes inconsistent CR status with the actual state.

#### Solution:
Does not treat progressing error as a fatal error

#### References:
https://pkg.go.dev/k8s.io/api/apps/v1#DeploymentSpec

### What type of PR is it?
[Bug Fix]

### Todos

### What is the Jira issue?
https://issues.apache.org/jira/projects/SUBMARINE/issues/SUBMARINE-1012

### How should this be tested?

### Screenshots (if appropriate)

### Questions:
* Do the license files need updating? No
* Are there breaking changes for older versions? No
* Does this need new documentation? No

Author: MortalHappiness <b07901069@ntu.edu.tw>

Signed-off-by: Kevin <pingsutw@apache.org>

Closes #744 from MortalHappiness/SUBMARINE-1012 and squashes the following commits:

34e94cc9 [MortalHappiness] SUBMARINE-1012. Submarine CR status is inconsistent with actual state