SUBMARINE-1215. Fix java.lang.NullPointerException in K8sSubmitter.getInfo() method
### What is this PR for?
When there is no mlflow, a `NullPointerException` will appear in the background. This PR is mainly to fix the background problem.

### What type of PR is it?
Bug Fix

### Todos
* [x] - Use Optional to avoid NullPointerException

### What is the Jira issue?
https://issues.apache.org/jira/browse/SUBMARINE-1215

### How should this be tested?
After fixing the background problem, the button will always turn when there is no mlflow in the k8s. I don't know if the front also needs to deal with it.
![image](https://user-images.githubusercontent.com/12069428/158007459-2ade1263-fc43-4a32-a991-d79c1925737b.png)

### Screenshots (if appropriate)
No

### Questions:
* Do the license files need updating? No
* Are there breaking changes for older versions? No
* Does this need new documentation? No

Author: cdmikechen <cdmikechen@hotmail.com>

Signed-off-by: Kevin <pingsutw@apache.org>

Closes #898 from cdmikechen/SUBMARINE-1215 and squashes the following commits:

be42ae04 [cdmikechen] SUBMARINE-1215. Fix java.lang.NullPointerException in K8sSubmitter.getInfo() method