SUBMARINE-1204. Notebook names cannot start with "submarine"
### What is this PR for?
The name of the notebook cannot start with "submarine". Otherwise, the service name will conflict with the other services that the operator created.
For example, if the name of the notebook is "submarine-server", then it will conflict with the service "submarine-server" that the operator created.
This PR aims to avoid this problem by adding the notebookId in front of notebookName at k8sSubmitter, but the notebookId will not display on frontend.

### What type of PR is it?
[Bug Fix]

### Todos

### What is the Jira issue?
https://issues.apache.org/jira/projects/SUBMARINE/issues/SUBMARINE-1204?filter=allopenissues
### How should this be tested?
Run the submarine service, and check `kubectl get notebook -n submarine-user-test`, if the notebook name in k8s is an <id>-<name> format, and the frontend only displays notebook name, then it is correct.
### Screenshots (if appropriate)

### Questions:
* Do the license files need updating? No
* Are there breaking changes for older versions? No
* Does this need new documentation? No

Author: Richie FY <b08901039@ntu.edu.tw>

Signed-off-by: Chi-Sheng Liu <chishengliu@apache.org>

Closes #922 from FYRichie/SUBMARINE-1204 and squashes the following commits:

94c8cf0a [Richie FY] remove comment
83887546 [Richie FY] add id to notebook api test
7151b769 [Richie FY] findNotebook add notebook id
f92a2112 [Richie FY] findNotebook list all notebook for match
d95c9827 [Richie FY] findNotebook add notebook id
3d9f9ad2 [Richie FY] SUBMARINE-1204. Notebook names cannot start with submarine