SUBMARINE-1174. K8sSubmitter refactoring involving Notebook create and delete method
### What is this PR for?
The current K8sSubmitter contains a lot of CRD processing logic in addition to the implementation of the interface.
The purpose of this PR is to split the CRD processing logic from the class as far as possible and merge it into the resource object for processing.

The current PR refactored part of the code so that we can see the changes of the code after refactoring.

### What type of PR is it?
Refactoring

### Todos
* [x] - Refactoring `K8sSubmitter`  involving Notebook create and delete method (Configmap/PVC/IstioVirtualService/NoteBookCR)
* [x] - Add k8s mock client/server for test

### What is the Jira issue?
https://issues.apache.org/jira/browse/SUBMARINE-1174

### How should this be tested?
The refactored codes can use the original test case.

### Screenshots (if appropriate)
No

### Questions:
* Do the license files need updating? No
* Are there breaking changes for older versions? No
* Does this need new documentation? No

Author: cdmikechen <cdmikechen@hotmail.com>

Signed-off-by: Kevin <pingsutw@apache.org>

Closes #870 from cdmikechen/SUBMARINE-1174 and squashes the following commits:

b30fdea6 [cdmikechen] rebase SUBMARINE-1276 and handling null pointer exceptions
ecf1d2b6 [cdmikechen] SUBMARINE-1174. K8sSubmitter refactoring involving Notebook create and delete method