SUBMARINE-1379. Fix a bug where tasks from missing k8s resources could not be deleted on the UI
### What is this PR for?
If the backend CR is manually deleted, the submarine-server is not currently checking if the specified resource still exists, resulting in a situation where the experiment cannot be deleted in UI.

### What type of PR is it?
Bug Fix

### Todos
* [x] - Remove `throwsApiException` in Notebook, Experiment and Serve.
* [x] - Fixed a bug where deleting a record on the UI, the entire page data clear

### What is the Jira issue?
https://issues.apache.org/jira/browse/SUBMARINE-1379

### How should this be tested?
CI tests can cover code, but they can only represent processes that handle normal conditions.
I removed the method of throwing an 404 exception error (`.throwsApiException()`) when the resource does not exist, which results in the return of the status object.
In this way, we can return the result normally, even without this resource

### Screenshots (if appropriate)
NA

### Questions:
* Do the license files need updating? No
* Are there breaking changes for older versions? No
* Does this need new documentation? No

Author: cdmikechen <cdmikechen@apache.org>

Signed-off-by: cdmikechen <cdmikechen@apache.org>

Closes #1073 from cdmikechen/SUBMARINE-1379 and squashes the following commits:

03a70110 [cdmikechen] fix list delete error
ec5c8b58 [cdmikechen] fix serve delete
3f829b41 [cdmikechen] fix notebook delete
b58271e4 [cdmikechen] fix experiment delete